small changes to get toolbar

+++ source/android/AndroidTouchScope/app/src/main/java/de/uni_weimar/mheinz/androidtouchscope/WaveData.java
@@ -3,24 +3,24 @@ package de.uni_weimar.mheinz.androidtouchscope;
 public class WaveData
 {
     public byte[] data;
-    public double voltageScale;
-    public double voltageOffset;
+    public float voltageScale;
+    public float voltageOffset;
 
-    public double timeScale;
-    public double timeOffset;
+    public float timeScale;
+    public float timeOffset;
 
     public WaveData()
     {
         data = null;
-        voltageScale = 0.0;
-        voltageOffset = 0.0;
-        timeScale = 0.0;
-        timeOffset = 0.0;
+        voltageScale = 0.0f;
+        voltageOffset = 0.0f;
+        timeScale = 0.0f;
+        timeOffset = 0.0f;
     }
 
     public WaveData(byte[] data,
-                    double voltageScale, double voltageOffset,
-                    double timeScale,    double timeOffset)
+                    float voltageScale, float voltageOffset,
+                    float timeScale,    float timeOffset)
     {
         this.data = data;
         this.voltageScale = voltageScale;
